Address comments in PR
Move test to test_slurm.py

Add link to SchedMD bug in test description

Add variable for max_retries
