+++ tests/integration-tests/tests/schedulers/test_slurm.py
@@ -483,6 +483,80 @@ def test_scontrol_reboot(
     )
 
 
+@pytest.mark.usefixtures("region", "os", "instance", "scheduler")
+def test_update_slurm_reconfigure_race_condition(
+    pcluster_config_reader,
+    clusters_factory,
+    test_datadir,
+    scheduler_commands_factory,
+):
+    """
+    Test race condition between restart of slurmctld and scontrol reconfigure.
+
+    In Slurm 21.08 it looks like cloud nodes may not get powered-down after their
+    SuspendTime has expired if a cluster update is performed, which restarts the
+    slurmctld daemon and immediately performs an scontrol reconfigure.
+
+    See https://bugs.schedmd.com/show_bug.cgi?id=13953
+    """
+
+    max_count_cr1 = 10
+    scale_down_idle_time_mins = 5
+
+    cluster_config = pcluster_config_reader(
+        config_file="pcluster.config.yaml",
+        output_file="pcluster.config.initial.yaml",
+        max_count_cr1=max_count_cr1,
+        scale_down_idle_time_mins=scale_down_idle_time_mins,
+    )
+    cluster = clusters_factory(cluster_config)
+    remote_command_executor = RemoteCommandExecutor(cluster)
+    slurm_commands = scheduler_commands_factory(remote_command_executor)
+
+    assert_compute_node_states(slurm_commands, compute_nodes=None, expected_states=["idle~"])
+
+    max_retries = 5
+    for iter in range(1, max_retries + 1):
+
+        job_id_1 = slurm_commands.submit_command_and_assert_job_accepted(
+            submit_command_args={
+                "nodes": 2,
+                "slots": 2,
+                "command": "srun sleep 300",
+                "raise_on_error": False,
+            }
+        )
+        slurm_commands.wait_job_running(job_id_1)
+        nodelist = slurm_commands.get_job_info(job_id_1, field="NodeList")
+        nodes = remote_command_executor.run_remote_command(
+            f"sinfo -N --nodes {nodelist} -h -O NodeHost:100 | sort | uniq"
+        ).stdout.splitlines()
+        nodes = [node.strip() for node in nodes]
+        slurm_commands.cancel_job(job_id_1)
+
+        max_count_cr1 = max_count_cr1 + 3
+
+        updated_config_file = pcluster_config_reader(
+            config_file="pcluster.config.yaml",
+            output_file=f"pcluster.config.iter_{iter}.yaml",
+            max_count_cr1=max_count_cr1,
+            scale_down_idle_time_mins=scale_down_idle_time_mins,
+        )
+        cluster.update(
+            config_file=updated_config_file,
+            wait=True,
+        )
+
+        # Check that nodes get powered down by Slurm.
+        retry(wait_fixed=seconds(30), stop_max_delay=minutes(2 * scale_down_idle_time_mins))(
+            assert_compute_node_states
+        )(
+            scheduler_commands=slurm_commands,
+            compute_nodes=nodes,
+            expected_states=["idle%"],
+        )
+
+
 def _assert_cluster_initial_conditions(
     scheduler_commands, expected_num_dummy, expected_num_instance_node, expected_num_static
 ):
